ショップ　ログイン関係、data.sql

+++ tasukuru/src/main/java/com/tasukuru/repository/ShopsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.tasukuru.entity.Shop;
 
 
-public interface ShopsRepository extends JpaRepository<Shop, Long> {
+public interface ShopsRepository extends JpaRepository<Shop, Integer> {
 //	List<Shop> findByNameContaining(String keyword);
 }
